来店情報がない場合の処理を追加

+++ edit.php
@@ -159,29 +159,31 @@ if (isset($_POST['UPDATE'])) {
     }
 
     // 来店情報の編集
-    $visit_sql = "UPDATE 来店記録 SET 日時=:date, 人数=:number, 続柄=:relation, メニュー=:menu WHERE 来店ID=:vid AND 顧客ID=:id";
-    $visit = $dbo->prepare($visit_sql);
-    $visit_del_sql = "DELETE FROM 来店記録 WHERE 来店ID=:vid AND 顧客ID=:id";
-    $visit_del = $dbo->prepare($visit_del_sql);
+    if (isset($forms['VISITID'])) {
+        $visit_sql = "UPDATE 来店記録 SET 日時=:date, 人数=:number, 続柄=:relation, メニュー=:menu WHERE 来店ID=:vid AND 顧客ID=:id";
+        $visit = $dbo->prepare($visit_sql);
+        $visit_del_sql = "DELETE FROM 来店記録 WHERE 来店ID=:vid AND 顧客ID=:id";
+        $visit_del = $dbo->prepare($visit_del_sql);
 
-    $i = 0;
-    foreach ($forms['VISITID'] as $vid) {
-        if (empty($forms['DATE'])) {
-            // 来店日時が空っぽなら削除する
-            $visit_del->bindParam(':vid', $vid);
-            $visit_del->bindParam(':id', $forms['ID']);
-            $res = $visit_del->execute();
-        } else {
-            // 更新処理
-            $visit->bindParam(':vid', $vid);
-            $visit->bindParam(':id', $forms['ID']);
-            $visit->bindParam(':date', $forms['DATE'][$i]);
-            $visit->bindParam(':number', $forms['NUMBER'][$i]);
-            $visit->bindParam(':relation', $forms['RELATION'][$i]);
-            $visit->bindParam(':menu', $forms['MENU'][$i]);
-            $res = $visit->execute();
+        $i = 0;
+        foreach ($forms['VISITID'] as $vid) {
+            if (empty($forms['DATE'][$i])) {
+                // 来店日時が空っぽなら削除する
+                $visit_del->bindParam(':vid', $vid);
+                $visit_del->bindParam(':id', $forms['ID']);
+                $res = $visit_del->execute();
+            } else {
+                // 更新処理
+                $visit->bindParam(':vid', $vid);
+                $visit->bindParam(':id', $forms['ID']);
+                $visit->bindParam(':date', $forms['DATE'][$i]);
+                $visit->bindParam(':number', $forms['NUMBER'][$i]);
+                $visit->bindParam(':relation', $forms['RELATION'][$i]);
+                $visit->bindParam(':menu', $forms['MENU'][$i]);
+                $res = $visit->execute();
+            }
+            $i++;
         }
-        $i++;
     }
 }
 
@@ -283,13 +285,16 @@ $client_visit = $res->fetchAll(PDO::FETCH_ASSOC);
     </div>
 
     <div id="visit">
+    <?php if (empty($client_visit)) { ?>
+        <p class="none">来店情報は登録されていません。</p>
+    <?php } else { ?>
         来店情報<br>
         <p style="font-size: 0.8em;">※日時を消すと削除します。</p>
-    <?php foreach ($client_visit as $visit) { ?>
+        <?php foreach ($client_visit as $visit) { ?>
         <hr><br>
         <input value="<?= $visit['来店ID'] ?>" type="hidden" name="VISITID[]">
         <div id="date">
-            <label><span>日時：</span><input value="<?= $visit['日時'] ?>" type="text" name="DATE[]" size="16" maxlength="40" required="required"></label>
+            <label><span>日時：</span><input value="<?= $visit['日時'] ?>" type="text" name="DATE[]" size="16" maxlength="40"></label>
         </div>
         <div id="number">
             <label><span>人数：</span><input value="<?= $visit['人数'] ?>" type="text" name="NUMBER[]" size="3" maxlength="3" required="required"></label>
@@ -300,6 +305,7 @@ $client_visit = $res->fetchAll(PDO::FETCH_ASSOC);
         <div id="menu">
             <label><span>メニュー：</span><textarea name="MENU[]"><?= $visit['メニュー'] ?></textarea></label>
         </div>
+        <?php } ?>
     <?php } ?>
     </div>
 </fieldset>
